Factory of SURVIVAL objects with Gompertz distributions

Creates a SURVIVAL object with an Gompertz distribution.

Usage

s_gompertz(...)

Arguments

...

Parameters to define the distribution. See the Parameters for details

Value

a SURVIVAL object of the Gompertz distribution family. See the documentation of s_factory for the methods available for SURVIVAL objects

Parameters

To create an exponential survival object the following options are available:

scale and shape to specify the canonical parameter of the distribution, or

surv, t and shape for the proportion surviving (no events) at time t and shape, or

fail and t and shape for the proportion failing (events) at time t and shape.

scale = -log(surv)·shape/(exp(shape·t))

scale = -log(1-fail)·shape/(exp(shape·t))

The parameters should be spell correctly as partial matching is not available

Examples

s_gompertz(scale = 1, shape = 1.5)
#> SURVIVAL object
#> Distribution:  GOMPERTZ 
#> scale : 1 
#> shape : 1.5 
s_gompertz(surv = 0.4, t = 2, shape = 1.5)
#> SURVIVAL object
#> Distribution:  GOMPERTZ 
#> scale : 0.07201454 
#> shape : 1.5 
s_gompertz(fail = 0.6, t = 2, shape = 1.5)
#> SURVIVAL object
#> Distribution:  GOMPERTZ 
#> scale : 0.07201454 
#> shape : 1.5